break actions and state into separate files
May just reference the actions in the components they are used later
